requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
pytesseract>=0.3.10
Pillow>=10.0.0
mss>=9.0.1
pyautogui>=0.9.54
keyboard>=0.13.5
beautifulsoup4>=4.12.0
PyQt5>=5.15.0
